+++ src/base/b-tree/CHANGELOG.md
@@ -14,13 +14,13 @@ Changelog
 #### :house: Internal
 
 * Added new methods `siblings`, `children`, `fold`, `unfold`
-* Added mod `clickableArea`
+* Added a new modifier `clickableArea`
 
 #### :bug: Bug Fix
 
 * Fixed passing props to nested trees
 * Fixed an issue with the prop `itemProps` not being added to items attributes
-* Fixed adding `folded_false` class to items without children
+* Fixed adding the `folded_false` class to items without children
 
 ## v3.0.0-rc.164 (2021-03-22)